+++ src/components/templates/Sidebar.jsx
@@ -1,8 +1,8 @@
 import { useEffect, useState } from "react";
-import { Link, useNavigate } from "react-router-dom";
 import styled from "styled-components";
 import TweetForm from "../organisms/TweetForm";
-import axios from "../../utils/axios"; // カスタムインスタンスをインポート
+import WithDrawal from "../organisms/WithDrawal";
+import SidebarStyledLink from "../atoms/SidebarStyledLink";
 
 const SidebarContainer = styled.div`
   flex: 1;
@@ -11,22 +11,6 @@ const SidebarContainer = styled.div`
   flex-direction: column;
   gap: 10px;
   margin-right: 50px;
-`;
-
-const StyledLink = styled(Link)`
-  text-decoration: none;
-  display: block;
-  padding: 10px;
-  background-color: #1da1f2;
-  color: white;
-  border: none;
-  border-radius: 20px;
-  cursor: pointer;
-  text-align: center;
-
-  &:hover {
-    background-color: #0d8ed9;
-  }
 `;
 
 // ユーザーIDをlocalStorageから取得する関数
@@ -45,43 +29,20 @@ const useUserId = () => {
 
 const Sidebar = () => {
   const userId = useUserId(); // ログインユーザーのIDを取得
-  const navigate = useNavigate();
-  const [isDeleted, setIsDeleted] = useState(false); // ユーザーが退会したかどうかを表すstate
-
-  // 退会APIを呼び出す関数
-  const deleteUser = async () => {
-    try {
-      const response = await axios.delete(`/users/delete`);
-      if (response.status === 200) {
-        alert("退会に成功しました");
-        // ユーザーをログアウト
-        localStorage.removeItem("id");
-        localStorage.removeItem("token");
-        // ユーザーが退会したことを表すstateを更新
-        setIsDeleted(true);
-      }
-    } catch (error) {
-      console.error("退会に失敗しました:", error);
-    }
-  };
-
-  // ユーザーが退会したときに画面遷移を行う
-  useEffect(() => {
-    if (isDeleted) {
-      navigate("/");
-      window.location.reload();
-    }
-  }, [isDeleted, navigate]);
 
   return (
     <SidebarContainer>
-      <StyledLink to="/">ホーム</StyledLink>
-      <StyledLink to="/groups">メッセージ</StyledLink>
-      <StyledLink to="/notifications">通知</StyledLink>
-      <StyledLink>設定</StyledLink>
-      <StyledLink>ログアウト</StyledLink>
-      {userId && <StyledLink to={`/user/${userId}`}>プロフィール</StyledLink>}
-      <StyledLink onClick={() => deleteUser(userId)}>退会</StyledLink>
+      <SidebarStyledLink to="/">ホーム</SidebarStyledLink>
+      <SidebarStyledLink to="/groups">メッセージ</SidebarStyledLink>
+      <SidebarStyledLink to="/notifications">通知</SidebarStyledLink>
+      <SidebarStyledLink>設定</SidebarStyledLink>
+      <SidebarStyledLink>ログアウト</SidebarStyledLink>
+      {userId && (
+        <SidebarStyledLink to={`/user/${userId}`}>
+          プロフィール
+        </SidebarStyledLink>
+      )}
+      <WithDrawal userId={userId} />
       <TweetForm />
     </SidebarContainer>
   );
